antok: More functions and clean-up.
Fixed the problems with the Initializer now by rewriting parts of the
configuration file and the part in the Initializer concerned with reading it.
Also added the Mass function, so all functions in functions.hpp can now be used
in the yaml file. Verified that the results are the same with both new and old
version of the treereader.

Also removed quite some commented debug output and restructured a little bit to
make things more readable. The initializeEvent() method has become quite a
beast, will probably have to be broken down at some point, especially
considering that more functions will have to be implemented.

+++ src/event.cxx
@@ -24,13 +24,14 @@ antok::Event::Event() {
 
 void antok::Event::update(antok::Data& data) {
 
-	const int& PION_MASS = antok::Constants::charged_pion_mass();
+	const double& PION_MASS = antok::Constants::charged_pion_mass();
 
 	rawData = &data;
 
 	_pSum.SetXYZT(0., 0., 0., 0.);
 	for(unsigned int i = 0; i < _p.size(); ++i) {
 		_p.at(i).SetXYZM(data.Mom_x.at(i), data.Mom_y.at(i), data.Mom_z.at(i), PION_MASS);
+//		_p.at(i).Print();
 		_pSum += _p.at(i);
 	}
